+++ e310x/src/gpio0/drive.rs
@@ -295,162 +295,162 @@ impl R {
 impl W {
     #[doc = "Bit 0"]
     #[inline(always)]
-    pub fn pin0(&mut self) -> Pin0W<DriveSpec> {
+    pub fn pin0(&mut self) -> Pin0W<'_, DriveSpec> {
         Pin0W::new(self, 0)
     }
     #[doc = "Bit 1"]
     #[inline(always)]
-    pub fn pin1(&mut self) -> Pin1W<DriveSpec> {
+    pub fn pin1(&mut self) -> Pin1W<'_, DriveSpec> {
         Pin1W::new(self, 1)
     }
     #[doc = "Bit 2"]
     #[inline(always)]
-    pub fn pin2(&mut self) -> Pin2W<DriveSpec> {
+    pub fn pin2(&mut self) -> Pin2W<'_, DriveSpec> {
         Pin2W::new(self, 2)
     }
     #[doc = "Bit 3"]
     #[inline(always)]
-    pub fn pin3(&mut self) -> Pin3W<DriveSpec> {
+    pub fn pin3(&mut self) -> Pin3W<'_, DriveSpec> {
         Pin3W::new(self, 3)
     }
     #[doc = "Bit 4"]
     #[inline(always)]
-    pub fn pin4(&mut self) -> Pin4W<DriveSpec> {
+    pub fn pin4(&mut self) -> Pin4W<'_, DriveSpec> {
         Pin4W::new(self, 4)
     }
     #[doc = "Bit 5"]
     #[inline(always)]
-    pub fn pin5(&mut self) -> Pin5W<DriveSpec> {
+    pub fn pin5(&mut self) -> Pin5W<'_, DriveSpec> {
         Pin5W::new(self, 5)
     }
     #[doc = "Bit 6"]
     #[inline(always)]
-    pub fn pin6(&mut self) -> Pin6W<DriveSpec> {
+    pub fn pin6(&mut self) -> Pin6W<'_, DriveSpec> {
         Pin6W::new(self, 6)
     }
     #[doc = "Bit 7"]
     #[inline(always)]
-    pub fn pin7(&mut self) -> Pin7W<DriveSpec> {
+    pub fn pin7(&mut self) -> Pin7W<'_, DriveSpec> {
         Pin7W::new(self, 7)
     }
     #[doc = "Bit 8"]
     #[inline(always)]
-    pub fn pin8(&mut self) -> Pin8W<DriveSpec> {
+    pub fn pin8(&mut self) -> Pin8W<'_, DriveSpec> {
         Pin8W::new(self, 8)
     }
     #[doc = "Bit 9"]
     #[inline(always)]
-    pub fn pin9(&mut self) -> Pin9W<DriveSpec> {
+    pub fn pin9(&mut self) -> Pin9W<'_, DriveSpec> {
         Pin9W::new(self, 9)
     }
     #[doc = "Bit 10"]
     #[inline(always)]
-    pub fn pin10(&mut self) -> Pin10W<DriveSpec> {
+    pub fn pin10(&mut self) -> Pin10W<'_, DriveSpec> {
         Pin10W::new(self, 10)
     }
     #[doc = "Bit 11"]
     #[inline(always)]
-    pub fn pin11(&mut self) -> Pin11W<DriveSpec> {
+    pub fn pin11(&mut self) -> Pin11W<'_, DriveSpec> {
         Pin11W::new(self, 11)
     }
     #[doc = "Bit 12"]
     #[inline(always)]
-    pub fn pin12(&mut self) -> Pin12W<DriveSpec> {
+    pub fn pin12(&mut self) -> Pin12W<'_, DriveSpec> {
         Pin12W::new(self, 12)
     }
     #[doc = "Bit 13"]
     #[inline(always)]
-    pub fn pin13(&mut self) -> Pin13W<DriveSpec> {
+    pub fn pin13(&mut self) -> Pin13W<'_, DriveSpec> {
         Pin13W::new(self, 13)
     }
     #[doc = "Bit 14"]
     #[inline(always)]
-    pub fn pin14(&mut self) -> Pin14W<DriveSpec> {
+    pub fn pin14(&mut self) -> Pin14W<'_, DriveSpec> {
         Pin14W::new(self, 14)
     }
     #[doc = "Bit 15"]
     #[inline(always)]
-    pub fn pin15(&mut self) -> Pin15W<DriveSpec> {
+    pub fn pin15(&mut self) -> Pin15W<'_, DriveSpec> {
         Pin15W::new(self, 15)
     }
     #[doc = "Bit 16"]
     #[inline(always)]
-    pub fn pin16(&mut self) -> Pin16W<DriveSpec> {
+    pub fn pin16(&mut self) -> Pin16W<'_, DriveSpec> {
         Pin16W::new(self, 16)
     }
     #[doc = "Bit 17"]
     #[inline(always)]
-    pub fn pin17(&mut self) -> Pin17W<DriveSpec> {
+    pub fn pin17(&mut self) -> Pin17W<'_, DriveSpec> {
         Pin17W::new(self, 17)
     }
     #[doc = "Bit 18"]
     #[inline(always)]
-    pub fn pin18(&mut self) -> Pin18W<DriveSpec> {
+    pub fn pin18(&mut self) -> Pin18W<'_, DriveSpec> {
         Pin18W::new(self, 18)
     }
     #[doc = "Bit 19"]
     #[inline(always)]
-    pub fn pin19(&mut self) -> Pin19W<DriveSpec> {
+    pub fn pin19(&mut self) -> Pin19W<'_, DriveSpec> {
         Pin19W::new(self, 19)
     }
     #[doc = "Bit 20"]
     #[inline(always)]
-    pub fn pin20(&mut self) -> Pin20W<DriveSpec> {
+    pub fn pin20(&mut self) -> Pin20W<'_, DriveSpec> {
         Pin20W::new(self, 20)
     }
     #[doc = "Bit 21"]
     #[inline(always)]
-    pub fn pin21(&mut self) -> Pin21W<DriveSpec> {
+    pub fn pin21(&mut self) -> Pin21W<'_, DriveSpec> {
         Pin21W::new(self, 21)
     }
     #[doc = "Bit 22"]
     #[inline(always)]
-    pub fn pin22(&mut self) -> Pin22W<DriveSpec> {
+    pub fn pin22(&mut self) -> Pin22W<'_, DriveSpec> {
         Pin22W::new(self, 22)
     }
     #[doc = "Bit 23"]
     #[inline(always)]
-    pub fn pin23(&mut self) -> Pin23W<DriveSpec> {
+    pub fn pin23(&mut self) -> Pin23W<'_, DriveSpec> {
         Pin23W::new(self, 23)
     }
     #[doc = "Bit 24"]
     #[inline(always)]
-    pub fn pin24(&mut self) -> Pin24W<DriveSpec> {
+    pub fn pin24(&mut self) -> Pin24W<'_, DriveSpec> {
         Pin24W::new(self, 24)
     }
     #[doc = "Bit 25"]
     #[inline(always)]
-    pub fn pin25(&mut self) -> Pin25W<DriveSpec> {
+    pub fn pin25(&mut self) -> Pin25W<'_, DriveSpec> {
         Pin25W::new(self, 25)
     }
     #[doc = "Bit 26"]
     #[inline(always)]
-    pub fn pin26(&mut self) -> Pin26W<DriveSpec> {
+    pub fn pin26(&mut self) -> Pin26W<'_, DriveSpec> {
         Pin26W::new(self, 26)
     }
     #[doc = "Bit 27"]
     #[inline(always)]
-    pub fn pin27(&mut self) -> Pin27W<DriveSpec> {
+    pub fn pin27(&mut self) -> Pin27W<'_, DriveSpec> {
         Pin27W::new(self, 27)
     }
     #[doc = "Bit 28"]
     #[inline(always)]
-    pub fn pin28(&mut self) -> Pin28W<DriveSpec> {
+    pub fn pin28(&mut self) -> Pin28W<'_, DriveSpec> {
         Pin28W::new(self, 28)
     }
     #[doc = "Bit 29"]
     #[inline(always)]
-    pub fn pin29(&mut self) -> Pin29W<DriveSpec> {
+    pub fn pin29(&mut self) -> Pin29W<'_, DriveSpec> {
         Pin29W::new(self, 29)
     }
     #[doc = "Bit 30"]
     #[inline(always)]
-    pub fn pin30(&mut self) -> Pin30W<DriveSpec> {
+    pub fn pin30(&mut self) -> Pin30W<'_, DriveSpec> {
         Pin30W::new(self, 30)
     }
     #[doc = "Bit 31"]
     #[inline(always)]
-    pub fn pin31(&mut self) -> Pin31W<DriveSpec> {
+    pub fn pin31(&mut self) -> Pin31W<'_, DriveSpec> {
         Pin31W::new(self, 31)
     }
 }