+++ e310x/src/i2c0/ctr.rs
@@ -25,12 +25,12 @@ impl R {
 impl W {
     #[doc = "Bit 6 - I2C core interrupt enable bit"]
     #[inline(always)]
-    pub fn ien(&mut self) -> IenW<CtrSpec> {
+    pub fn ien(&mut self) -> IenW<'_, CtrSpec> {
         IenW::new(self, 6)
     }
     #[doc = "Bit 7 - I2C core enable bit"]
     #[inline(always)]
-    pub fn en(&mut self) -> EnW<CtrSpec> {
+    pub fn en(&mut self) -> EnW<'_, CtrSpec> {
         EnW::new(self, 7)
     }
 }