+++ e310x/src/i2c0/prer_hi.rs
@@ -16,7 +16,7 @@ impl R {
 impl W {
     #[doc = "Bits 0:7"]
     #[inline(always)]
-    pub fn value(&mut self) -> ValueW<PrerHiSpec> {
+    pub fn value(&mut self) -> ValueW<'_, PrerHiSpec> {
         ValueW::new(self, 0)
     }
 }