+++ hifive1-examples/examples/sh_rgb_blink.rs
@@ -6,7 +6,7 @@
 
 use hifive1::{
     clock,
-    hal::{e310x::CLINT, prelude::*, DeviceResources},
+    hal::{prelude::*, DeviceResources},
     pins, Led,
 };
 use semihosting::{println, process::exit};
@@ -14,11 +14,12 @@ use semihosting::{println, process::exit};
 #[riscv_rt::entry]
 fn main() -> ! {
     let dr = DeviceResources::take().unwrap();
+    let cp = dr.core_peripherals;
     let p = dr.peripherals;
     let pins = dr.pins;
 
     // Configure clocks
-    let clocks = clock::configure(p.PRCI, p.AONCLK, 320.mhz().into());
+    clock::configure(p.PRCI, p.AONCLK, 320.mhz().into());
 
     // get all 3 led pins in a tuple (each pin is it's own type here)
     let rgb_pins = pins!(pins, (led_red, led_green, led_blue));
@@ -26,8 +27,8 @@ fn main() -> ! {
     // get leds as the Led trait in an array so we can index them
     let mut ileds: [&mut dyn Led; 3] = [&mut tleds.0, &mut tleds.1, &mut tleds.2];
 
-    // Get the sleep struct from CLINT
-    let mut sleep = CLINT::delay();
+    // Get the MTIMER peripheral from CLINT
+    let mut mtimer = cp.clint.mtimer();
 
     const N_TOGGLES: usize = 4;
     const STEP: u32 = 500; // 500ms
@@ -37,7 +38,7 @@ fn main() -> ! {
         for (i, led) in ileds.iter_mut().enumerate() {
             led.toggle().unwrap();
             println!("LED {i} toggled. New state: {}", led.is_on());
-            sleep.delay_ms(STEP);
+            mtimer.delay_ms(STEP);
         }
     }
     println!("Done toggling LEDs");